--- packages/api-gateway/package.json
@@ -1,30 +1,0 @@
-{
-	"name": "@siiges-services/api-gateway",
-	"version": "1.0.0",
-	"description": "This package is for manage the rutes of the api",
-	"author": "cetherdev <robert.hct@gmail.com>",
-	"homepage": "https://github.com/robertotorres-dev/siiges-services#readme",
-	"license": "ISC",
-	"main": "src/index.js",
-	"directories": {
-		"src": "src",
-		"test": "__tests__"
-	},
-	"files": [
-		"src",
-		"__tests__"
-	],
-	"scripts": {
-		"lint": "eslint .",
-		"lint:fix": "eslint --fix .",
-		"test": "jest",
-		"test:watch": "jest --watch",
-		"precommit": "yarn lint && yarn jest" 
-	},
-	"bugs": {
-		"url": "https://github.com/robertotorres-dev/siiges-services/issues"
-	},
-	"dependencies": {
-		"@siiges-services/authentication": "^1.0.0"
-	}
-}
