--- packages/api-gateway/src/index.js
@@ -1,5 +1,0 @@
-function apiGateway() {
-  // TODO
-}
-
-module.exports = apiGateway;
