+++ src/jest.config.js
