+++ src/packages/api-gateway/package.json
@@ -15,30 +15,16 @@
 		"__tests__"
 	],
 	"scripts": {
-		"start:dev": "nodemon src/index.js",
-		"inspect": "node --inspect src/index.js",
-		"test": "jest --runInBand",
-		"test:watch": "jest --watch",
 		"lint": "eslint .",
-		"lint:fix": "eslint . --fix"
+		"lint:fix": "eslint --fix .",
+		"test": "jest",
+		"test:watch": "jest --watch",
+		"precommit": "yarn lint && yarn jest" 
 	},
 	"bugs": {
 		"url": "https://github.com/robertotorres-dev/siiges-services/issues"
 	},
 	"dependencies": {
-		"@fastify/autoload": "^5.0.0",
-		"@fastify/cors": "^8.0.0",
-		"@siiges-services/authentication": "^1.0.0",
-		"@siiges-services/usuario": "^1.0.0",
-		"fastify": "^4.2.0"
-	},
-	"devDependencies": {
-		"@hapi/boom": "^9.1.4",
-		"eslint": "^8.12.0",
-		"eslint-config-airbnb-base": "^15.0.0",
-		"eslint-plugin-import": "^2.25.2",
-		"jest": "^27.5.1",
-		"jest-mongoose-mock": "^1.2.0",
-		"nodemon": "^2.0.15"
+		"@siiges-services/authentication": "^1.0.0"
 	}
 }
