+++ src/packages/api-gateway/src/index.js
@@ -1,6 +1,5 @@
-// Internal dependencies
-const HttpServer = require('./drivers/http/server');
+function apiGateway() {
+  // TODO
+}
 
-HttpServer.start().catch((err) =>
-	console.error('Something went wrong when running server', err)
-);
+module.exports = apiGateway;
