+++ src/packages/authentication/__tests__/authentication.test.js
@@ -1,7 +1,11 @@
-'use strict';
+const addition = require('../src/test');
 
-const authentication = require('..');
-
-describe('authentication', () => {
-    it('needs tests');
+describe('given a set of numbers', () => {
+  const adden2 = 4;
+  const adden1 = 5;
+  describe('when we call addition function', () => {
+    test('then adition should return its sum', () => {
+      expect(addition(adden1, adden2)).toEqual(9);
+    });
+  });
 });
